+++ power_law_tool_equ.hpp
@@ -12,11 +12,11 @@
 #include <tuple>
 
 
-std::vector<unsigned long long> generate_power_law_point_process(double alpha,
+std::vector<unsigned int> generate_power_law_point_process(double alpha,
                                                     double tau_0 = 1.0,
                                                     int sample_amount = 100000
                                                     );
 
-std::tuple<double, double, std::vector<int>, std::vector<double>> dfa(std::vector<unsigned long long> RW_list, double alpha = 0.0, int first_l = 4, int t_last_l= 10000);
+std::tuple<double, double, std::vector<int>, std::vector<double>> dfa(std::vector<unsigned int> RW_list, double alpha = 0.0, int first_l = 4, int t_last_l= 10000);
 
 #endif
